refactor: modify app structure
